Network.AWS.CloudHSMv2.CreateCluster

Description

Creates a new AWS CloudHSM cluster.

Creating a Request

createCluster

Arguments

:: NonEmpty Text
ccSubnetIds

-> Text
ccHSMType

-> CreateCluster

Creates a value of CreateCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccSourceBackupId - The identifier (ID) of the cluster backup to restore. Use this value to restore the cluster from a backup instead of creating a new cluster. To find the backup ID, use DescribeBackups .

ccSubnetIds - The identifiers (IDs) of the subnets where you are creating the cluster. You must specify at least one subnet. If you specify multiple subnets, they must meet the following criteria: * All subnets must be in the same virtual private cloud (VPC). * You can specify only one subnet per Availability Zone.

ccHSMType - The type of HSM to use in the cluster. Currently the only allowed value is hsm1.medium .

data CreateCluster

See: createCluster smart constructor.

Request Lenses

ccSourceBackupId :: Lens' CreateCluster (Maybe Text)

The identifier (ID) of the cluster backup to restore. Use this value to restore the cluster from a backup instead of creating a new cluster. To find the backup ID, use DescribeBackups .

ccSubnetIds :: Lens' CreateCluster (NonEmpty Text)

The identifiers (IDs) of the subnets where you are creating the cluster. You must specify at least one subnet. If you specify multiple subnets, they must meet the following criteria: * All subnets must be in the same virtual private cloud (VPC). * You can specify only one subnet per Availability Zone.

ccHSMType :: Lens' CreateCluster Text

The type of HSM to use in the cluster. Currently the only allowed value is hsm1.medium .

Destructuring the Response

createClusterResponse

Arguments

:: Int
ccrsResponseStatus

-> CreateClusterResponse

Creates a value of CreateClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccrsCluster - Information about the cluster that was created.

ccrsResponseStatus - -- | The response status code.

data CreateClusterResponse

See: createClusterResponse smart constructor.

Response Lenses

ccrsCluster :: Lens' CreateClusterResponse (Maybe Cluster)

Information about the cluster that was created.

ccrsResponseStatus :: Lens' CreateClusterResponse Int

- | The response status code.