Network.AWS.CloudHSMv2.CreateHSM

Description

Creates a new hardware security module (HSM) in the specified AWS CloudHSM cluster.

Creating a Request

createHSM

Arguments

:: Text
chClusterId

-> Text
chAvailabilityZone

-> CreateHSM

Creates a value of CreateHSM with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chIPAddress - The HSM's IP address. If you specify an IP address, use an available address from the subnet that maps to the Availability Zone where you are creating the HSM. If you don't specify an IP address, one is chosen for you from that subnet.

chClusterId - The identifier (ID) of the HSM's cluster. To find the cluster ID, use DescribeClusters .

chAvailabilityZone - The Availability Zone where you are creating the HSM. To find the cluster's Availability Zones, use DescribeClusters .

data CreateHSM

See: createHSM smart constructor.

Request Lenses

chIPAddress :: Lens' CreateHSM (Maybe Text)

The HSM's IP address. If you specify an IP address, use an available address from the subnet that maps to the Availability Zone where you are creating the HSM. If you don't specify an IP address, one is chosen for you from that subnet.

chClusterId :: Lens' CreateHSM Text

The identifier (ID) of the HSM's cluster. To find the cluster ID, use DescribeClusters .

chAvailabilityZone :: Lens' CreateHSM Text

The Availability Zone where you are creating the HSM. To find the cluster's Availability Zones, use DescribeClusters .

Destructuring the Response

createHSMResponse

Arguments

:: Int
chrsResponseStatus

-> CreateHSMResponse

Creates a value of CreateHSMResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chrsHSM - Information about the HSM that was created.

chrsResponseStatus - -- | The response status code.

data CreateHSMResponse

See: createHSMResponse smart constructor.

Response Lenses

chrsHSM :: Lens' CreateHSMResponse (Maybe HSM)

Information about the HSM that was created.

chrsResponseStatus :: Lens' CreateHSMResponse Int

- | The response status code.