Network.AWS.CloudHSMv2.DescribeBackups

Description

Gets information about backups of AWS CloudHSM clusters.

This is a paginated operation, which means that each response might contain only a subset of all the backups. When the response contains only a subset of backups, it includes a NextToken value. Use this value in a subsequent DescribeBackups request to get more backups. When you receive a response with no NextToken (or an empty or null value), that means there are no more backups to get.

Creating a Request

describeBackups :: DescribeBackups

Creates a value of DescribeBackups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbFilters - One or more filters to limit the items returned in the response. Use the backupIds filter to return only the specified backups. Specify backups by their backup identifier (ID). Use the clusterIds filter to return only the backups for the specified clusters. Specify clusters by their cluster identifier (ID). Use the states filter to return only backups that match the specified state.

dbNextToken - The NextToken value that you received in the previous response. Use this value to get more backups.

dbMaxResults - The maximum number of backups to return in the response. When there are more backups than the number you specify, the response contains a NextToken value.

data DescribeBackups

See: describeBackups smart constructor.

Request Lenses

dbFilters :: Lens' DescribeBackups (HashMap Text [Text])

One or more filters to limit the items returned in the response. Use the backupIds filter to return only the specified backups. Specify backups by their backup identifier (ID). Use the clusterIds filter to return only the backups for the specified clusters. Specify clusters by their cluster identifier (ID). Use the states filter to return only backups that match the specified state.

dbNextToken :: Lens' DescribeBackups (Maybe Text)

The NextToken value that you received in the previous response. Use this value to get more backups.

dbMaxResults :: Lens' DescribeBackups (Maybe Natural)

The maximum number of backups to return in the response. When there are more backups than the number you specify, the response contains a NextToken value.

Destructuring the Response

describeBackupsResponse

Arguments

:: Int
dbrsResponseStatus

-> DescribeBackupsResponse

Creates a value of DescribeBackupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbrsBackups - A list of backups.

dbrsNextToken - An opaque string that indicates that the response contains only a subset of backups. Use this value in a subsequent DescribeBackups request to get more backups.

dbrsResponseStatus - -- | The response status code.

data DescribeBackupsResponse

See: describeBackupsResponse smart constructor.

Response Lenses

dbrsBackups :: Lens' DescribeBackupsResponse [Backup]

A list of backups.

dbrsNextToken :: Lens' DescribeBackupsResponse (Maybe Text)

An opaque string that indicates that the response contains only a subset of backups. Use this value in a subsequent DescribeBackups request to get more backups.

dbrsResponseStatus :: Lens' DescribeBackupsResponse Int

- | The response status code.