Network.AWS.CloudHSMv2.DescribeClusters

Description

Gets information about AWS CloudHSM clusters.

This is a paginated operation, which means that each response might contain only a subset of all the clusters. When the response contains only a subset of clusters, it includes a NextToken value. Use this value in a subsequent DescribeClusters request to get more clusters. When you receive a response with no NextToken (or an empty or null value), that means there are no more clusters to get.

Creating a Request

describeClusters :: DescribeClusters

Creates a value of DescribeClusters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcFilters - One or more filters to limit the items returned in the response. Use the clusterIds filter to return only the specified clusters. Specify clusters by their cluster identifier (ID). Use the vpcIds filter to return only the clusters in the specified virtual private clouds (VPCs). Specify VPCs by their VPC identifier (ID). Use the states filter to return only clusters that match the specified state.

dcNextToken - The NextToken value that you received in the previous response. Use this value to get more clusters.

dcMaxResults - The maximum number of clusters to return in the response. When there are more clusters than the number you specify, the response contains a NextToken value.

data DescribeClusters

See: describeClusters smart constructor.

Request Lenses

dcFilters :: Lens' DescribeClusters (HashMap Text [Text])

One or more filters to limit the items returned in the response. Use the clusterIds filter to return only the specified clusters. Specify clusters by their cluster identifier (ID). Use the vpcIds filter to return only the clusters in the specified virtual private clouds (VPCs). Specify VPCs by their VPC identifier (ID). Use the states filter to return only clusters that match the specified state.

dcNextToken :: Lens' DescribeClusters (Maybe Text)

The NextToken value that you received in the previous response. Use this value to get more clusters.

dcMaxResults :: Lens' DescribeClusters (Maybe Natural)

The maximum number of clusters to return in the response. When there are more clusters than the number you specify, the response contains a NextToken value.

Destructuring the Response

describeClustersResponse

Arguments

:: Int
dcrsResponseStatus

-> DescribeClustersResponse

Creates a value of DescribeClustersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrsNextToken - An opaque string that indicates that the response contains only a subset of clusters. Use this value in a subsequent DescribeClusters request to get more clusters.

dcrsClusters - A list of clusters.

dcrsResponseStatus - -- | The response status code.

data DescribeClustersResponse

See: describeClustersResponse smart constructor.

Response Lenses

dcrsNextToken :: Lens' DescribeClustersResponse (Maybe Text)

An opaque string that indicates that the response contains only a subset of clusters. Use this value in a subsequent DescribeClusters request to get more clusters.

dcrsClusters :: Lens' DescribeClustersResponse [Cluster]

A list of clusters.

dcrsResponseStatus :: Lens' DescribeClustersResponse Int

- | The response status code.