Network.AWS.CloudHSMv2.ListTags

Description

Gets a list of tags for the specified AWS CloudHSM cluster.

This is a paginated operation, which means that each response might contain only a subset of all the tags. When the response contains only a subset of tags, it includes a NextToken value. Use this value in a subsequent ListTags request to get more tags. When you receive a response with no NextToken (or an empty or null value), that means there are no more tags to get.

Creating a Request

listTags

Arguments

:: Text
ltResourceId

-> ListTags

Creates a value of ListTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltNextToken - The NextToken value that you received in the previous response. Use this value to get more tags.

ltMaxResults - The maximum number of tags to return in the response. When there are more tags than the number you specify, the response contains a NextToken value.

ltResourceId - The cluster identifier (ID) for the cluster whose tags you are getting. To find the cluster ID, use DescribeClusters .

data ListTags

See: listTags smart constructor.

Request Lenses

ltNextToken :: Lens' ListTags (Maybe Text)

The NextToken value that you received in the previous response. Use this value to get more tags.

ltMaxResults :: Lens' ListTags (Maybe Natural)

The maximum number of tags to return in the response. When there are more tags than the number you specify, the response contains a NextToken value.

ltResourceId :: Lens' ListTags Text

The cluster identifier (ID) for the cluster whose tags you are getting. To find the cluster ID, use DescribeClusters .

Destructuring the Response

listTagsResponse

Arguments

:: Int
ltrsResponseStatus

-> NonEmpty Tag
ltrsTagList

-> ListTagsResponse

Creates a value of ListTagsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltrsNextToken - An opaque string that indicates that the response contains only a subset of tags. Use this value in a subsequent ListTags request to get more tags.

ltrsResponseStatus - -- | The response status code.

ltrsTagList - A list of tags.

data ListTagsResponse

See: listTagsResponse smart constructor.

Response Lenses

ltrsNextToken :: Lens' ListTagsResponse (Maybe Text)

An opaque string that indicates that the response contains only a subset of tags. Use this value in a subsequent ListTags request to get more tags.

ltrsResponseStatus :: Lens' ListTagsResponse Int

- | The response status code.

ltrsTagList :: Lens' ListTagsResponse (NonEmpty Tag)

A list of tags.