Network.AWS.CloudHSMv2.Types

Description

Service Configuration

cloudHSMv2 :: Service

API version 2017-04-28 of the Amazon CloudHSM V2 SDK configuration.

Errors

_CloudHSMInternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because of an AWS CloudHSM internal failure. The request can be retried.

_CloudHSMServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because an error occurred.

_CloudHSMInvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it is not a valid request.

_CloudHSMAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the requester does not have permission to perform the requested operation.

_CloudHSMResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it refers to a resource that cannot be found.

BackupPolicy

data BackupPolicy

Constructors

Default

BackupState

data BackupState

Constructors

BSCreateInProgress

BSDeleted

BSReady

ClusterState

data ClusterState

Constructors

Active

CreateInProgress

Degraded

DeleteInProgress

Deleted

InitializeInProgress

Initialized

Uninitialized

UpdateInProgress

HSMState

data HSMState

Constructors

HSActive

HSCreateInProgress

HSDegraded

HSDeleteInProgress

HSDeleted

Backup

data Backup

Contains information about a backup of an AWS CloudHSM cluster.

See: backup smart constructor.

backup

Arguments

:: Text
bBackupId

-> Backup

Creates a value of Backup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bClusterId - The identifier (ID) of the cluster that was backed up.

bCreateTimestamp - The date and time when the backup was created.

bBackupState - The state of the backup.

bBackupId - The identifier (ID) of the backup.

bClusterId :: Lens' Backup (Maybe Text)

The identifier (ID) of the cluster that was backed up.

bCreateTimestamp :: Lens' Backup (Maybe UTCTime)

The date and time when the backup was created.

bBackupState :: Lens' Backup (Maybe BackupState)

The state of the backup.

bBackupId :: Lens' Backup Text

The identifier (ID) of the backup.

Certificates

data Certificates

Contains one or more certificates or a certificate signing request (CSR).

See: certificates smart constructor.

certificates :: Certificates

Creates a value of Certificates with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cManufacturerHardwareCertificate - The HSM hardware certificate issued (signed) by the hardware manufacturer.

cClusterCSR - The cluster's certificate signing request (CSR). The CSR exists only when the cluster's state is UNINITIALIZED .

cHSMCertificate - The HSM certificate issued (signed) by the HSM hardware.

cClusterCertificate - The cluster certificate issued (signed) by the issuing certificate authority (CA) of the cluster's owner.

cAWSHardwareCertificate - The HSM hardware certificate issued (signed) by AWS CloudHSM.

cManufacturerHardwareCertificate :: Lens' Certificates (Maybe Text)

The HSM hardware certificate issued (signed) by the hardware manufacturer.

cClusterCSR :: Lens' Certificates (Maybe Text)

The cluster's certificate signing request (CSR). The CSR exists only when the cluster's state is UNINITIALIZED .

cHSMCertificate :: Lens' Certificates (Maybe Text)

The HSM certificate issued (signed) by the HSM hardware.

cClusterCertificate :: Lens' Certificates (Maybe Text)

The cluster certificate issued (signed) by the issuing certificate authority (CA) of the cluster's owner.

cAWSHardwareCertificate :: Lens' Certificates (Maybe Text)

The HSM hardware certificate issued (signed) by AWS CloudHSM.

Cluster

data Cluster

Contains information about an AWS CloudHSM cluster.

See: cluster smart constructor.

cluster :: Cluster

Creates a value of Cluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cPreCoPassword - The default password for the cluster's Pre-Crypto Officer (PRECO) user.

cStateMessage - A description of the cluster's state.

cState - The cluster's state.

cSubnetMapping - A map of the cluster's subnets and their corresponding Availability Zones.

cHSMs - Contains information about the HSMs in the cluster.

cVPCId - The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

cSourceBackupId - The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

cCertificates - Contains one or more certificates or a certificate signing request (CSR).

cSecurityGroup - The identifier (ID) of the cluster's security group.

cClusterId - The cluster's identifier (ID).

cCreateTimestamp - The date and time when the cluster was created.

cBackupPolicy - The cluster's backup policy.

cHSMType - The type of HSM that the cluster contains.

cPreCoPassword :: Lens' Cluster (Maybe Text)

The default password for the cluster's Pre-Crypto Officer (PRECO) user.

cStateMessage :: Lens' Cluster (Maybe Text)

A description of the cluster's state.

cState :: Lens' Cluster (Maybe ClusterState)

The cluster's state.

cSubnetMapping :: Lens' Cluster (HashMap Text Text)

A map of the cluster's subnets and their corresponding Availability Zones.

cHSMs :: Lens' Cluster [HSM]

Contains information about the HSMs in the cluster.

cVPCId :: Lens' Cluster (Maybe Text)

The identifier (ID) of the virtual private cloud (VPC) that contains the cluster.

cSourceBackupId :: Lens' Cluster (Maybe Text)

The identifier (ID) of the backup used to create the cluster. This value exists only when the cluster was created from a backup.

cCertificates :: Lens' Cluster (Maybe Certificates)

Contains one or more certificates or a certificate signing request (CSR).

cSecurityGroup :: Lens' Cluster (Maybe Text)

The identifier (ID) of the cluster's security group.

cClusterId :: Lens' Cluster (Maybe Text)

The cluster's identifier (ID).

cCreateTimestamp :: Lens' Cluster (Maybe UTCTime)

The date and time when the cluster was created.

cBackupPolicy :: Lens' Cluster (Maybe BackupPolicy)

The cluster's backup policy.

cHSMType :: Lens' Cluster (Maybe Text)

The type of HSM that the cluster contains.

HSM

data HSM

Contains information about a hardware security module (HSM) in an AWS CloudHSM cluster.

See: hsm smart constructor.

hsm

Arguments

:: Text
hsmHSMId

-> HSM

Creates a value of HSM with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

hsmStateMessage - A description of the HSM's state.

hsmState - The HSM's state.

hsmEniId - The identifier (ID) of the HSM's elastic network interface (ENI).

hsmSubnetId - The subnet that contains the HSM's elastic network interface (ENI).

hsmAvailabilityZone - The Availability Zone that contains the HSM.

hsmClusterId - The identifier (ID) of the cluster that contains the HSM.

hsmEniIP - The IP address of the HSM's elastic network interface (ENI).

hsmHSMId - The HSM's identifier (ID).

hsmStateMessage :: Lens' HSM (Maybe Text)

A description of the HSM's state.

hsmState :: Lens' HSM (Maybe HSMState)

The HSM's state.

hsmEniId :: Lens' HSM (Maybe Text)

The identifier (ID) of the HSM's elastic network interface (ENI).

hsmSubnetId :: Lens' HSM (Maybe Text)

The subnet that contains the HSM's elastic network interface (ENI).

hsmAvailabilityZone :: Lens' HSM (Maybe Text)

The Availability Zone that contains the HSM.

hsmClusterId :: Lens' HSM (Maybe Text)

The identifier (ID) of the cluster that contains the HSM.

hsmEniIP :: Lens' HSM (Maybe Text)

The IP address of the HSM's elastic network interface (ENI).

hsmHSMId :: Lens' HSM Text

The HSM's identifier (ID).

Tag

data Tag

Contains a tag. A tag is a key-value pair.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - The key of the tag.

tagValue - The value of the tag.

tagKey :: Lens' Tag Text

The key of the tag.

tagValue :: Lens' Tag Text

The value of the tag.